x/truststore/types/message_update_attestation_test.go
package types

import (
	"testing"

	"github.com/blarsy/truststore/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgUpdateAttestation_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdateAttestation
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdateAttestation{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdateAttestation{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
